+++ src/main/java/be/ugent/idlab/knows/functions/agent/functionModelProvider/fno/NAMESPACES.java
@@ -10,6 +10,7 @@ public enum NAMESPACES {
     DOAP("http://usefulinc.com/ns/doap#"),
     FNO("https://w3id.org/function/ontology#"),
     FNOI("https://w3id.org/function/vocabulary/implementation#"),
+    FNOC("https://w3id.org/function/vocabulary/composition#"),
     FNOM("https://w3id.org/function/vocabulary/mapping#"),
     LIB("http://example.com/library#"),
     RDF("http://www.w3.org/1999/02/22-rdf-syntax-ns#"),