+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-07-12
+### Added
+- `FnoFunctionModelProvider`: Can now parse Function Composition.
+- `Agent`: Can execute Function Compositions.
+
+### Fixed
+- `Instantiator`: fixed Windows path resolving issues.
+- `Instantiator`: fixed URL decoding for special characters in path.
+
 ## [0.1.0] - 2022-05-25
 
 ### Added